Replace isz usage with uint64

+++ plugins/opengl_rendering/src/opengl_shader.cpp
@@ -1,4 +1,5 @@
 #include "opengl_shader.h"
+#include "tbx/common/int.h"
 #include "tbx/debugging/macros.h"
 #include <glad/glad.h>
 #include <glm/gtc/type_ptr.hpp>
@@ -28,7 +29,7 @@ namespace tbx::plugins::openglrendering
     {
         GLint length = 0;
         glGetShaderiv(shader_id, GL_INFO_LOG_LENGTH, &length);
-        std::string error_log(static_cast<std::size_t>(length), '\0');
+        std::string error_log(static_cast<uint64>(length), '\0');
         glGetShaderInfoLog(shader_id, length, &length, error_log.data());
         TBX_ASSERT(
             false,
@@ -41,7 +42,7 @@ namespace tbx::plugins::openglrendering
     {
         GLint length = 0;
         glGetProgramiv(program_id, GL_INFO_LOG_LENGTH, &length);
-        std::string error_log(static_cast<std::size_t>(length), '\0');
+        std::string error_log(static_cast<uint64>(length), '\0');
         glGetProgramInfoLog(program_id, length, &length, error_log.data());
         TBX_ASSERT(
             false,
